requirements.txt
scrapy
scrapy-splash
pylint==1.9.3
shub==2.8.1
google-cloud-storage==1.13.0
pyasn1-modules==0.2.1
requests==2.18.4
futures==3.2.0
scrapinghub==2.0.3
langdetect==1.0.7
beautifulsoup4==4.6.3
python-dateutil==2.7.5
html2text==2018.1.9
markdown==3.0.1
js2py==0.60
